test(lib): run the analyzer regime on the test project, warning-free
ReScene.Tests now builds with EnableNETAnalyzers + AnalysisLevel=latest-All +
EnforceCodeStyleInBuild (+ GenerateDocumentationFile so IDE0005 runs), matching
the library and ending the inversion where only production code was analyzed.

Test-only conventions get justified .editorconfig exemptions (CA1707/IDE1006
underscore test names, CA5394 seeded Random, CA1003 Action-based recorder
hooks, IDE0330 net8.0 has no System.Threading.Lock). Everything else was
fixed for real: IDE0055/0004/0230/0042/0059/0290/0370 via per-rule dotnet
format passes, plus hand fixes for CA1861 constant-array args, CA1823 unused
field, CA1304/CA1311 ToLowerInvariant, CS1573/CS1574 doc crefs, IDE0060
unused parameter — and one CA1307 (ordinal IndexOf) the sweep surfaced in
SRRFile.ExtractStoredFiles itself.

Both TFM legs verified: 1558 tests green, zero build warnings.

+++ ReScene.Tests/SyntheticSampleBuilder.cs
@@ -56,7 +56,7 @@ internal static class SyntheticSampleBuilder
         // Calculate total sizes
         uint hdrlSize = (uint)(4 + hdrlBytes.Length); // "hdrl" + children
         uint moviSize = (uint)(4 + moviBytes.Length); // "movi" + children
-        uint riffSize = (uint)(4 + 8 + hdrlSize + 8 + moviSize); // "AVI " + LIST headers
+        uint riffSize = 4 + 8 + hdrlSize + 8 + moviSize; // "AVI " + LIST headers
 
         // Write RIFF header
         bw.Write(Encoding.ASCII.GetBytes("RIFF"));
@@ -89,6 +89,7 @@ internal static class SyntheticSampleBuilder
     /// its sample/movie files share identical track-2 bytes; the writer/parser
     /// suites use the default 42.
     /// </param>
+    /// <param name="path">Destination path for the built MKV.</param>
     public static string BuildMKV(string path, int block2Seed = 42)
     {
         using var ms = new MemoryStream();
